refactor: rename common service package to application
