+++ src/main/java/com/handwoong/everyonewaiter/common/application/port/TimeHolder.java
@@ -1,0 +1,6 @@
+package com.handwoong.everyonewaiter.common.application.port;
+
+public interface TimeHolder {
+
+    long millis();
+}
