--- src/main/java/com/handwoong/everyonewaiter/common/service/port/TimeHolder.java
@@ -1,6 +1,0 @@
-package com.handwoong.everyonewaiter.common.service.port;
-
-public interface TimeHolder {
-
-    long millis();
-}
